fix: 각 페이지 별 성격에 맞도록 페이지 명칭 변경 (#28)
