+++ app/contact/page.tsx
@@ -1,9 +1,8 @@
 import ContactPageContent from '@/app/contact/ContactPageContent';
 
 export const metadata = {
-  title: '더 알아보기',
+  title: '문의하기',
 };
-
 export default function ContactPage() {
   return <ContactPageContent />;
 }
