--- app/report/page.tsx
@@ -1,8 +1,0 @@
-import ReportPageContent from '@/app/report/ReportPageContent';
-
-export const metadata = {
-  title: '문의하기',
-};
-export default function ReportPage() {
-  return <ReportPageContent />;
-}
